requirements.txt
Django
gunicorn
django-heroku
django-allauth
django-measurement
django-crispy-forms
m26